seperated postdata into 2 files
